Additional tests for geo module + refactoring in tests
